backend/requirements.txt
# LifePilot Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# JWT Authentication
PyJWT==2.8.0

# HTTP Client (for external APIs)
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2
